pyproject.toml
[tool.poetry]
name = "goals-zone"
version = "0.1.0"
description = "gz.meneses.pt"
authors = ["meneses-pt <andre@meneses.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-webhook = "^1.3.1"
Django = "^5.0.4"
django4_background_tasks = "^1.2.9"
django_admin_rangefilter = "^0.12.4"
django_hosts = "^6.0"
djangorestframework = "^3.15.1"
fake_headers = "^1.0.2"
geoip2 = "^4.8.0"
lxml = "^5.2.1"
Markdown = "^3.4.1"
proxy_list = "^0.3.8"
psycopg = "^3.1.18"
pytz = "^2024.1"
requests = "^2.31.0"
slack_webhook = "^1.0.7"
spacy = "3.7.4"
tweepy = "^4.14"
Pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
gunicorn = {extras = ["gevent"], version = "^22.0.0"}
Brotli = "^1.0.9"
free-proxy = "^1.1"
django-cors-headers = "^4.3.1"
pymemcache = "^4.0.0"
retry = "^0.9.2"
black = "^24.3.0"
ruff = "^0.4.0"
scrapfly-sdk = {extras = ["all"], version = "^0.8.16"}
asyncio = "^3.4.3"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.ruff]
output-format = "full"
line-length = 120
exclude = [".venv", ".git", "helm", "terraform", "build"]
target-version = "py38"
lint.select = ["A", "ANN", "ASYNC", "B", "C4", "E", "F", "I", "ISC", "N", "Q", "SIM", "T20", "UP", "W"]
lint.fixable = ["ANN", "C4", "I"]
lint.ignore = [
    "ANN101"
]
